Attempted to update map based on navigator.geolocation
