+++ client/src/map.ts
@@ -1,22 +1,8 @@
-export const getLocation = () => {
-    let options = {
-        enableHighAccuracy: true,
-        timeout: 5000,
-        maximumAge: 0
-    };
-    let coords: any
-    const success = (pos: any) => {
-        coords = pos.coords;
-    }
-    const error = (err: any) => {
-        console.warn(`ERROR(${err.code}): ${err.message}`);
-    }
-
-    if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(success,error,options);
-        return coords
-    }
-    else{
-        console.log("Geolocation is not supported by this browser.");
+export class LongLat {
+    lat: number
+    lng: number
+    constructor(lat: number, lng: number) {
+        this.lat = lat
+        this.lng = lng
     }
 }
